refactor: Repackage code for Aqua scanner
