+++ pkg/aqua/scanner/converter.go
@@ -1,4 +1,4 @@
-package aqua
+package scanner
 
 import (
 	"encoding/json"
