+++ pkg/aqua/scanner/converter_test.go
@@ -1,4 +1,4 @@
-package aqua
+package scanner
 
 import (
 	"os"
