+++ pkg/aqua/scanner/model.go
@@ -1,4 +1,4 @@
-package aqua
+package scanner
 
 type ResourceType int
