+++ pkg/aqua/scanner/scanner_test.go
@@ -1,4 +1,4 @@
-package aqua
+package scanner
 
 import (
 	"context"
