+++ pkg/aqua/scanner/test_fixtures/aqua_csp_report/with_vulnerabilities.json
