+++ pkg/aqua/scanner/test_fixtures/aqua_csp_report/without_vulnerabilities.json
